+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 1.0.0-0 / Unreleased
-  * .set() creates intermediate objects if missing
-  * .writeToForm() and .readFromForm() accept inputs with nested names
+  * [ADD] .set() creates intermediate objects if missing
+  * [ADD] .writeToForm() and .readFromForm() accept inputs with nested names
 
 # 0.0.1 / 2016-04-10
   * Initial release